src/backend/requirements.txt
django
djangorestframework
channels
psycopg2
redis
pyyaml